# finance-nestjs/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/payment-methods": {
      "get": {
        "operationId": "PaymentMethodController_find",
        "parameters": [
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Filtrar métodos de pagamento pelo nome",
            "schema": {
              "example": "Cartão de Crédito",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "payment-methods"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/payment-methods/{categoryId}": {
      "get": {
        "operationId": "PaymentMethodController_findOne",
        "parameters": [
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "description": "ID do método de pagamento a ser buscado",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "payment-methods"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/v1/files/{userId}/{filename}": {
      "get": {
        "operationId": "FileController_getFile",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "ID do usuário dono do arquivo",
            "schema": {
              "example": "123",
              "type": "string"
            }
          },
          {
            "name": "filename",
            "required": true,
            "in": "path",
            "description": "Nome do arquivo a ser recuperado",
            "schema": {
              "example": "foto.jpg",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "files"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/files/public/{directory}/{filename}": {
      "get": {
        "operationId": "FileController_getPublicFile",
        "parameters": [
          {
            "name": "directory",
            "required": true,
            "in": "path",
            "description": "Diretório público onde o arquivo está",
            "schema": {
              "example": "avatars",
              "type": "string"
            }
          },
          {
            "name": "filename",
            "required": true,
            "in": "path",
            "description": "Nome do arquivo público a ser recuperado",
            "schema": {
              "example": "avatar1.png",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "files"
        ]
      }
    },
    "/v1/issuers": {
      "get": {
        "operationId": "IssuerController_find",
        "parameters": [
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Filtra emissores pelo nome",
            "schema": {
              "example": "Banco do Brasil",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "issuers"
        ]
      }
    },
    "/v1/issuers/{id}": {
      "get": {
        "operationId": "IssuerController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do emissor a ser buscado",
            "schema": {
              "example": 123,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "issuers"
        ]
      }
    }
  },
  "info": {
    "title": "Finance APP",
    "description": "The finance API description",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Local environment"
    }
  ],
...
}